% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-data.R
\name{create.regr.data}
\alias{create.regr.data}
\title{A function for creating random regression data.}
\usage{
create.regr.data(n, p, ytype = "continuous")
}
\arguments{
\item{n}{The number of observations in the random data set.}

\item{p}{The number of parameters in the random data set. The function randomly chooses how many will be binary or continuous.}

\item{ytype}{Determines whether to have a binary or continuous response variable. Defaults to "continuous"}
}
\description{
This function returns a random
}
\examples{
create.regr.data(100,4,ytype="continuous")
create.regr.data(100,4,ytype="binary")
}
