% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{extended.glm}
\alias{extended.glm}
\title{A custom wrapper for base function glm(family=binomial("logit))}
\usage{
extended.glm(X, Y)
}
\arguments{
\item{X}{an n by p matrix of candidate predictors.}

\item{Y}{an n by 1 vector of binary responses}
}
\description{
This function takes a matrix of candidate predictors X and a vector of binary response variables Y. .
}
\examples{
extended.lm(X,Y)
}
