% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{extended.lasso}
\alias{extended.lasso}
\title{A custom wrapper for base function glmnet::glmnet(alpha=1).}
\usage{
extended.lasso(X, Y, lambda = NULL, ytype = "continuous")
}
\arguments{
\item{X}{an n by p matrix of candidate predictors.}

\item{Y}{an n by 1 vector of responses}

\item{lambda}{an optional lambda value. If no lambda is provided, an optimal lambda is
automatically determined using glmnet::cv.glmnet(alpha=1).}

\item{ytype}{description as to whether or not the response variable y is binary or continuous. Defaults to 'continuous'.}
}
\description{
This function takes a matrix of candidate predictors X and a vector of response variables Y and performs lasso regression.
}
\examples{
extended.lasso(X,Y)
}
