% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{extended.lm}
\alias{extended.lm}
\title{A custom wrapper for base function lm()}
\usage{
extended.lm(X, Y)
}
\arguments{
\item{X}{an n by p matrix of candidate predictors.}

\item{Y}{an n by 1 vector of responses}
}
\description{
This function takes a matrix of candidate predictors X and a vector of response variables Y.
}
\examples{
require(glmnet)
data("QuickStartExample")
x <- QuickStartExample$x
y <- QuickStartExample$y
extended.lm(x,y)
}
