% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{extended.randomlasso}
\alias{extended.randomlasso}
\title{An implementation of random lasso regression.}
\usage{
extended.randomlasso(
  X,
  Y,
  lambda = NULL,
  B = 200,
  q1 = length(Y),
  q2 = length(Y),
  ytype = "continuous"
)
}
\arguments{
\item{X}{an n by p matrix of candidate predictors.}

\item{Y}{an n by 1 vector of responses}

\item{lambda}{an optional lambda value. If no lambda is provided, an optimal lambda is automatically determined using glmnet::cv.glmnet(alpha=1).}

\item{B}{the number of Bootstrap samples. Default is 200.}

\item{q1}{the number of random predictors to use in step 1 of the algorithm. Default is length(Y). Must be less than or equal to this value.}

\item{q2}{the number of random predictors to use in step 2 of the algorithm. Default is length(Y). Must be less than or equal to this value.}

\item{ytype}{description as to whether or not the response variable y is binary or continuous. Defaults to 'continuous'.}
}
\description{
This function takes a matrix of candidate predictors X and a vector of response variables Y and performs random lasso regression.
}
\examples{
extended.randomlasso(X,Y)
}
