% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{extended.ridge}
\alias{extended.ridge}
\title{A custom wrapper for base function lm().}
\usage{
extended.ridge(X, Y, lambda = NULL, ytype = "continuous")
}
\arguments{
\item{X}{an n by p matrix of candidate predictors.}

\item{Y}{an n by 1 vector of responses}
}
\description{
This function takes a matrix of candidate predictors X and a vector of response variables Y.
}
\examples{
extended.lm(X,Y)
}
