% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-test-split.R
\name{train_test_split}
\alias{train_test_split}
\title{A function for splitting data into a train/test set}
\usage{
train_test_split(data, test_size = 0.2)
}
\arguments{
\item{data}{a data.frame or matrix where rows equal observations. Location of xs and ys does not matter.}

\item{test_size}{a real value in (0,1). Default is 0.2. If less than 0 or greater than 1 an error is returned.}
}
\description{
This function splits data into a training and testing set.
You can choose the test size.
It returns a list with the training data and test data as separate objects.
}
\examples{
df <- create.regr.data(100,4,ytype="continuous")$Data
train_test_slit(df)
}
