% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-data.R
\name{create.regr.data}
\alias{create.regr.data}
\title{A function for creating random regression data.}
\usage{
create.regr.data(n, p, ytype = "continuous", include.categorical = FALSE)
}
\arguments{
\item{n}{The number of observations in the random data set.}

\item{p}{The number of parameters in the random data set. The function randomly chooses how many will be binary or continuous.}

\item{ytype}{Determines whether to have a binary or continuous response variable. Defaults to "continuous"}

\item{include.categorical}{A boolean condition that determines whether or not to include categorical predictors. By default it is set to FALSE}
}
\description{
This function creates a random data.frame with n observations of p parameters.
You can choose whether or not the response variable y is binary or continuous.
It returns a list with the data.frame as well as a vector of the true beta values used to create the data.
}
\examples{
create.regr.data(100,4,ytype="continuous")
create.regr.data(100,4,ytype="binary")
}
