% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{extended.glm}
\alias{extended.glm}
\title{A custom wrapper for base function glm(family=binomial("logit))}
\usage{
extended.glm(X, Y)
}
\arguments{
\item{X}{an n by p matrix of candidate predictors.}

\item{Y}{an n by 1 vector of binary responses}
}
\description{
This function takes a matrix of candidate predictors X and a vector of binary response variables Y. .
}
\examples{
n <- 1000
B0 <- 3
B1 <- 5
X1 <- rnorm(n)
B2 <- -2
X2 <- rnorm(n)
X <- as.matrix(cbind(X1=X1,X2=X2))
Z <- B0 + B1*X1 + B2*X2
pr=exp(Z)/(1+exp(Z))
Y <- rbinom(n,1,pr)
fit <- extended.glm(X,Y2)
coef(fit)
}
